requirements.txt
datasets
sacrebleu 
py7zr
pandas
nltk
tqdm
PyYAML
matplotlib
notebook
boto3
mypy-boto3-s3
python-box==6.0.2
ensure==1.0.2
fastapi==0.78.0
uvicorn==0.18.3
Jinja2==3.1.2
torchsummary
pytorch-lightning==2.1.0
opencv-python==4.8.0.74
Pillow==10.0.0
scikit-image
tqdm
torchvision
scikit-image
pytorch_lightning
scipy


